Move every interfaces to src/Typings/index.d.ts
