+++ src/lib/help.ts
@@ -2,11 +2,12 @@ import Client from '../Client'
 import commands from './commands.json'
 import Utils from '../Utils'
 import responses from './responses.json'
+import { ICommandList } from '../Typings'
 
 export const help = (client: Client, command?: string): string => {
     if (command) {
         for (const catogary in commands) {
-            for (const index of (commands as commandList)[catogary]) {
+            for (const index of (commands as ICommandList)[catogary]) {
                 if (index.command === command) {
                     return `📚 *_Note: Add the Prefix infront of the command while using (${client._config.prefix})_*\n\n*📗 Command:* ${index.command}\n📙 *Description:* ${index.description}\n📘 *Usage:* ${client._config.prefix}${index.usage}`
                 }
@@ -15,7 +16,7 @@ export const help = (client: Client, command?: string): string => {
         return responses['invalid-command-short'].replace('{C}', command)
     }
     let base = `🤖 ${client._config.name} Command List 🤖\n\n💡 *Prefix ${client._config.prefix}*\n\n`
-    const cmds = commands as commandList
+    const cmds = commands as ICommandList
     const cats = Object.keys(cmds)
     for (const cat in cmds) {
         base += `*${Utils.capitalize(cat)}* ${Utils.emojies[cats.indexOf(cat)]}\n\`\`\``
@@ -26,13 +27,4 @@ export const help = (client: Client, command?: string): string => {
     }
     return `${base}📚 Use ${client._config.prefix}help <command_name> to view the full info. \n🔖 _Eg: ${client._config.prefix}help promote_`
 }
-export interface commandList {
-    [catogary: string]: command[]
-}
 
-export interface command {
-    command: string
-    description: string
-    usage: string
-    flags?: string[]
-}
